+++ Opera/APP/js/whatsNew.js
@@ -2,56 +2,20 @@
 
 // What's new in update
 newUpdate = {
-  msg: "New update!",
-  content: "<p>What's new in this update:</p><li>Changed layout, right now it's resizable</li><li>Changed style of options</li><li>Notifications won't 'attack' you on restart of your browser</li><li>Now you can disable/enable notifications for certain streamers</li><li>A lot of optimizations</li><p>Right now I'm pretty busy, so next update (1.5) will be done at summer</p>"
+  msg: "Hotfix",
+  content: "<p>What's new in this update:</p><li>Fixed background script</li><li>Fixed bug with zero followed channels</li><li>Added button in options 'in case of bugs'</li><br><p>My apologies for wrecking everything, if something not yet fixed you can contact me via Twitter <a href='https://twitter.com/iZarudny' target='_blank'>@iZarudny</a></p>"
 };
 
 // Changelog
 changes = [
-  "-1.4.5 a lot of UI improvements, script optimizations, added notifications settings for each streamer, resizable extension's window",
-  "-1.4.2 hotfix, fixed auth by twitch's account, fixed game links",
-  "-1.4.1 changed font to PermianSans Typeface, fixed ON/OFF statements, added notify on offline, optimized network code",
-  "-1.4 Fresh look, fixed 'Following List' page, various improvements",
-  "-1.3.9.9 Bug fixing",
-  "-1.3.9.8 Code optimization",
-  "-1.3.9.7 Fixed notifications",
-  "-1.3.9.6 Bugfix",
-  "-1.3.9.5 Hotfix",
-  "-1.3.9.4 Fixed zoomed preview ratio",
-  "-1.3.9.3 Fixed Donation tab, various improvements",
-  "-1.3.9.2 Added auth by TwitchTV account, fixed sound in notifications",
-  "-1.3.9.1 Bug fix",
-  "-1.3.9 Edited some animation, edited design, stability improvements",
-  "-1.3.8 Fixed bug with text on hover, added a new feature: zoomed preview",
-  "-1.3.7 Deleted 'Live update' script, various style editings",
-  "-1.3.6 Cumulative update",
-  "-1.3.5 Fixed update script",
-  "-1.3.4 Happy New Year!",
-  "-1.3.3 Cumulative update",
-  "-1.3.2 Style improvements",
-  "-1.3.1 Bug fixes, style improvements",
-  "-1.3.0 Added possibility to change style of online list, improvements",
-  "-1.2.9.9 Now options are easy to understand",
-  "-1.2.9 A little redesign, added DejaVu font",
-  "-1.2.8 Bug fixes, improvements",
-  "-1.2.7 Bug fixes",
-  "-1.2.6 Added 'Watch now!' to notifications",
-  "-1.2.5 Added Animation, thanks for Animate.css (http://daneden.me/animate)",
-  "-1.2.4 Reedited 'Live Update' script",
-  "-1.2.3 Bug fixes (3), eyep",
-  "-1.2.2 Bug fixes (2)",
-  "-1.2.0 Modified storage system",
-  "-1.1.9 Added version control for 'Live Update' script",
-  "-1.1.8 Bug fixes",
-  "-1.1.7 Added 'Live Update' script. Fixed minor bugs",
-  "-1.1.6 Bug fixes",
-  "-1.1.5 A little optimisation in style. Now you can disable duration of the stream. Button 'Found a bug?' now only viewable in options and in changes",
-  "-1.1.4 Fixed freezes on opening",
-  "-1.1.3 Bug with notifications, fixed stream duration",
-  "-1.1.2 Stream duration",
-  "-1.1.1 Hotfix",
-  "-1.1.0 Added sound effects for notifications. Added 'Version changes' page. And more...",
-  "-1.0.2 Resolved a problem which freezes app",
-  "-1.0.1 Bug fixes",
-  "-1.0.0 First publish in Google Web Store"
+  "-1.1.6 hotfix, fixed background script, fixed bug with zero followed channels",
+  "-1.1.5 a lot of UI improvements, script optimizations, added notifications settings for each streamer, resizable extension's window",
+  "-1.1.1 hotfix, fixed auth by twitch's account, fixed game links",
+  "-1.1 changed font to PermianSans Typeface, fixed ON/OFF statements, added notify on offline, optimized network code",
+  "-1.0 fresh look, added notifications (works only on 25+), fixed 'Following List' page, various improvements",
+  "-0.9.0.4 bug fixing",
+  "-0.9.0.3 code optimization",
+  "-0.9.0.2 fixed sound",
+  "-0.9.0.1 gugfix",
+  "-0.9.0.0 first publish in Opera Web Store"
 ];